Various ssize_t cleanups.
Primarily just avoids warning when converting between Py_ssize_t and ints.
Many cases the sizes are checked, but other cases just blindly cast - the
latter is done when it seems almost impossible the int will be large
enough to overflow, but that choice might seem arbitrary in some cases.

+++ com/win32com/src/PyIDispatch.cpp
@@ -110,7 +110,7 @@ PyObject *PyIDispatch::GetIDsOfNames(PyObject *self, PyObject *args)
     // @pyparm string|name||A name to query for
     // @pyparmalt1 [string, ...]|[name, ...]||A sequence of string names to query
     // @comm Currently the LCID can not be specified, and  LOCALE_SYSTEM_DEFAULT is used.
-    int argc = PyTuple_GET_SIZE(args);
+    int argc = (int)PyTuple_GET_SIZE(args);
     if (argc < 1)
         return PyErr_Format(PyExc_TypeError, "At least one argument must be supplied");
 
@@ -189,7 +189,7 @@ PyObject *PyIDispatch::GetIDsOfNames(PyObject *self, PyObject *args)
 static BOOL PyCom_MakeUntypedDISPPARAMS(PyObject *args, int numArgs, WORD wFlags, DISPPARAMS *pParm,
                                         PythonOleArgHelper **ppHelpers)
 {
-    int argc = PyObject_Length(args);
+    int argc = (int)PyObject_Length(args);
     DISPID dispidNamed = DISPID_PROPERTYPUT;
     // Clean initialize
     pParm->rgvarg = NULL;
@@ -261,7 +261,7 @@ PyObject *PyIDispatch::Invoke(PyObject *self, PyObject *args)
 #else
     PyErr_Clear();
 #endif
-    int argc = PyObject_Length(args);
+    int argc = (int)PyObject_Length(args);
     if (argc == -1)
         return NULL;
     if (argc < 4)
@@ -339,7 +339,7 @@ PyObject *PyIDispatch::InvokeTypes(PyObject *self, PyObject *args)
 {
     /* InvokeType(dispid, lcid, wflags, ELEMDESC resultType, ELEMDESC[] argTypes, arg1, arg2...) */
     PyErr_Clear();
-    int argc = PyObject_Length(args);
+    int argc = (int)PyObject_Length(args);
     if (argc == -1)
         return NULL;
     if (argc < 5)
@@ -362,7 +362,7 @@ PyObject *PyIDispatch::InvokeTypes(PyObject *self, PyObject *args)
     if (PyErr_Occurred())
         return NULL;
     int numArgs;
-    int argTypesLen = PyObject_Length(argsElemDescArray);
+    int argTypesLen = (int)PyObject_Length(argsElemDescArray);
     if (!PyTuple_Check(argsElemDescArray) || argTypesLen < argc - 5)
         return PyErr_Format(PyExc_TypeError,
                             "The array of argument types must be a tuple whose size is <= to the number of arguments.");
@@ -687,7 +687,7 @@ PyObject *PyIDispatchEx::InvokeEx(PyObject *self, PyObject *args)
 
     DISPPARAMS dispparams;
     PythonOleArgHelper *helpers;
-    if (!PyCom_MakeUntypedDISPPARAMS(invokeArgs, PyObject_Length(invokeArgs), flags, &dispparams, &helpers))
+    if (!PyCom_MakeUntypedDISPPARAMS(invokeArgs, (int)PyObject_Length(invokeArgs), flags, &dispparams, &helpers))
         return NULL;
 
     VARIANT varResult;
